refactor(Views v2): Refactor filters into separate class
