+++ src/Tribe/Views/V2/Rest_Endpoint.php
@@ -1,0 +1,13 @@
+<?php
+/**
+ *
+ *
+ * @package Tribe\Events\Views\V2
+ * @since TBD
+ */
+
+namespace Tribe\Events\Views\V2;
+
+
+class Rest_Endpoint {
+}
